apiserver/provisioner: Simplified a couple of tests, extracted ProvisioningInfo code and tests in separate files
